+++ assets/web-cart.lua
@@ -1,0 +1,12 @@
+function _init()
+    cls(1)
+    pset(64, 64, 8)
+end
+
+function _update()
+end
+
+function _draw()
+    pset(32, 32, 7)
+    pset(96, 96, 12)
+end